stiching the VIT-G agian
