+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.safetensors
 *.log
 *.json
+*.txt
