+++ PermSpec_Base.py
@@ -1,0 +1,15 @@
+from typing import NamedTuple
+from collections import defaultdict
+from re import L
+
+class PermutationSpec(NamedTuple):
+  perm_to_axes: dict
+  axes_to_perm: dict
+
+def permutation_spec_from_axes_to_perm(axes_to_perm: dict) -> PermutationSpec:
+  perm_to_axes = defaultdict(list)
+  for wk, axis_perms in axes_to_perm.items():
+    for axis, perm in enumerate(axis_perms):
+      if perm is not None:
+        perm_to_axes[perm].append((wk, axis))
+  return PermutationSpec(perm_to_axes=dict(perm_to_axes), axes_to_perm=axes_to_perm)
