+++ merge_PermSpec_SDXL.py
@@ -1,7 +1,7 @@
 # Ref: https://github.com/vladmandic/automatic/blob/dev/modules/merging/merge_PermSpec_SDXL.py
 # Tldr: Awful. So many layers are incorrect and missing!
 
-from weight_matching import PermutationSpec, permutation_spec_from_axes_to_perm
+from PermSpec_Base import PermutationSpec, permutation_spec_from_axes_to_perm
 
 def sdxl_permutation_spec() -> PermutationSpec:
     conv = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
@@ -1167,230 +1167,232 @@ def sdxl_permutation_spec() -> PermutationSpec:
 
             # Text Encoder 2
             # Should observe the original openCLIP Repo: https://github.com/openai/CLIP/blob/main/clip/model.py#L171
+            # Size / sequence has its pattern! Somehow not compatable with ViT-L!
+
             **dense("conditioner.embedders.1.model.token_embedding", None, None),
             **norm("conditioner.embedders.1.model.ln_final", "P_bg1430"),
             **skip("conditioner.embedders.1.model.logit_scale", None, None),
             **skip("conditioner.embedders.1.model.positional_embedding", None, None),
             **skip("conditioner.embedders.1.model.text_projection", None, None),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj", "P_bg1274", "P_bg1275", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj", "P_bg1276", "P_bg1277", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1274", "P_bg1275", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.0.ln_1", "P_bg1275"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1275", "P_bg1279", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1279", "P_bg1280", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.0.ln_2", "P_bg1280"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj", "P_bg1280", "P_bg1281", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj", "P_bg1282", "P_bg1283", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1280", "P_bg1281", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.1.ln_1", "P_bg1281"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1281", "P_bg1285", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1285", "P_bg1287", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.1.ln_2", "P_bg1287"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj", "P_bg1287", "P_bg1347", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj", "P_bg1289", "P_bg1290", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1287", "P_bg1347", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.2.ln_1", "P_bg1347"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1348", "P_bg1351", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1351", "P_bg1353", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.2.ln_2", "P_bg1353"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj", "P_bg1353", "P_bg1413", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj", "P_bg1355", "P_bg1356", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1353", "P_bg1413", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.3.ln_1", "P_bg1413"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1413", "P_bg1417", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1417", "P_bg1419", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.3.ln_2", "P_bg1419"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj", "P_bg1419", "P_bg1431", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj", "P_bg1421", "P_bg1422", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1419", "P_bg1431", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.4.ln_1", "P_bg1431"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1431", "P_bg1435", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1435", "P_bg1436", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.4.ln_2", "P_bg1436"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj", "P_bg1436", "P_bg1437", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj", "P_bg1438", "P_bg1439", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1436", "P_bg1437", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.5.ln_1", "P_bg1437"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1437", "P_bg1441", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1441", "P_bg1442", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.5.ln_2", "P_bg1442"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj", "P_bg1442", "P_bg1443", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj", "P_bg1444", "P_bg1445", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1442", "P_bg1443", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.6.ln_1", "P_bg1443"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1443", "P_bg1447", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1447", "P_bg1448", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.6.ln_2", "P_bg1448"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj", "P_bg1448", "P_bg1449", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj", "P_bg1450", "P_bg1451", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1448", "P_bg1449", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.7.ln_1", "P_bg1449"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1449", "P_bg1453", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1453", "P_bg1454", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.7.ln_2", "P_bg1454"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj", "P_bg1454", "P_bg1455", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj", "P_bg1456", "P_bg1457", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1454", "P_bg1455", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.8.ln_1", "P_bg1455"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1455", "P_bg1459", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1459", "P_bg1460", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.8.ln_2", "P_bg1460"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj", "P_bg1460", "P_bg1461", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj", "P_bg1462", "P_bg1463", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1460", "P_bg1461", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.9.ln_1", "P_bg1461"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1461", "P_bg1465", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1465", "P_bg1286", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.9.ln_2", "P_bg1286"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj", "P_bg1286", "P_bg1287", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj", "P_bg1288", "P_bg1466", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1286", "P_bg1287", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.10.ln_1", "P_bg1287"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1287", "P_bg1291", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1291", "P_bg1292", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.10.ln_2", "P_bg1292"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj", "P_bg1292", "P_bg1293", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj", "P_bg1294", "P_bg1295", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1292", "P_bg1293", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.11.ln_1", "P_bg1293"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1293", "P_bg1297", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1297", "P_bg1298", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.11.ln_2", "P_bg1298"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj", "P_bg1298", "P_bg1299", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj", "P_bg1300", "P_bg1301", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1298", "P_bg1299", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.12.ln_1", "P_bg1299"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1299", "P_bg1303", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1303", "P_bg1304", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.12.ln_2", "P_bg1304"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj", "P_bg1304", "P_bg1305", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj", "P_bg1306", "P_bg1307", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1304", "P_bg1305", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.13.ln_1", "P_bg1305"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1305", "P_bg1309", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1309", "P_bg1310", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.13.ln_2", "P_bg1310"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj", "P_bg1310", "P_bg1311", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj", "P_bg1312", "P_bg1313", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1310", "P_bg1311", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.14.ln_1", "P_bg1311"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1311", "P_bg1315", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1315", "P_bg1316", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.14.ln_2", "P_bg1316"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj", "P_bg1316", "P_bg1317", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj", "P_bg1318", "P_bg1319", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1316", "P_bg1317", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.15.ln_1", "P_bg1317"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1317", "P_bg1321", bias=True),
-            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1321", "P_bg1325", bias=True),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1321", "P_bg1322", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.15.ln_2", "P_bg1322"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj", "P_bg1322", "P_bg1323", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj", "P_bg1324", "P_bg1325", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1322", "P_bg1323", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.16.ln_1", "P_bg1323"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1323", "P_bg1327", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1327", "P_bg1328", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.16.ln_2", "P_bg1328"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj", "P_bg1328", "P_bg1329", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj", "P_bg1330", "P_bg1331", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1328", "P_bg1329", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.17.ln_1", "P_bg1329"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1329", "P_bg1333", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1333", "P_bg1334", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.17.ln_2", "P_bg1334"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj", "P_bg1334", "P_bg1335", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj", "P_bg1336", "P_bg1337", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1334", "P_bg1335", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.18.ln_1", "P_bg1335"),
-            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1339", "P_bg1339", bias=True),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1335", "P_bg1339", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1339", "P_bg1340", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.18.ln_2", "P_bg1340"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj", "P_bg1340", "P_bg1341", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj", "P_bg1342", "P_bg1343", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1340", "P_bg1341", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.19.ln_1", "P_bg1341"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1341", "P_bg1345", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1345", "P_bg1352", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.19.ln_2", "P_bg1352"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj", "P_bg1352", "P_bg1353", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj", "P_bg1354", "P_bg1346", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1352", "P_bg1353", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.20.ln_1", "P_bg1353"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1353", "P_bg1357", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1357", "P_bg1358", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.20.ln_2", "P_bg1358"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj", "P_bg1358", "P_bg1359", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj", "P_bg1360", "P_bg1361", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1358", "P_bg1359", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.21.ln_1", "P_bg1359"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1359", "P_bg1363", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1363", "P_bg1364", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.21.ln_2", "P_bg1364"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj", "P_bg1364", "P_bg1365", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj", "P_bg1366", "P_bg1367", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1364", "P_bg1365", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.22.ln_1", "P_bg1365"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1365", "P_bg1369", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1369", "P_bg1370", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.22.ln_2", "P_bg1370"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj", "P_bg1370", "P_bg1371", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj", "P_bg1372", "P_bg1373", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1370", "P_bg1371", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.23.ln_1", "P_bg1371"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1371", "P_bg1375", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1375", "P_bg1376", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.23.ln_2", "P_bg1376"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj", "P_bg1376", "P_bg1377", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj", "P_bg1378", "P_bg1379", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1376", "P_bg1377", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.24.ln_1", "P_bg1377"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1377", "P_bg1381", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1381", "P_bg1382", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.24.ln_2", "P_bg1382"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj", "P_bg1382", "P_bg1383", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj", "P_bg1384", "P_bg1385", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1382", "P_bg1383", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.25.ln_1", "P_bg1383"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1383", "P_bg1387", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1387", "P_bg1388", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.25.ln_2", "P_bg1388"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj", "P_bg1388", "P_bg1389", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj", "P_bg1390", "P_bg1391", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1388", "P_bg1389", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.26.ln_1", "P_bg1389"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1389", "P_bg1393", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1393", "P_bg1394", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.26.ln_2", "P_bg1394"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj", "P_bg1394", "P_bg1395", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj", "P_bg1396", "P_bg1397", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1394", "P_bg1395", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.27.ln_1", "P_bg1395"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1395", "P_bg1399", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1399", "P_bg1400", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.27.ln_2", "P_bg1400"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj", "P_bg1400", "P_bg1401", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj", "P_bg1402", "P_bg1403", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1400", "P_bg1401", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.28.ln_1", "P_bg1401"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1401", "P_bg1405", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1405", "P_bg1406", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.28.ln_2", "P_bg1406"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj", "P_bg1406", "P_bg1407", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj", "P_bg1408", "P_bg1409", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1406", "P_bg1407", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.29.ln_1", "P_bg1407"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1407", "P_bg1411", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1411", "P_bg1418", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.29.ln_2", "P_bg1418"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj", "P_bg1418", "P_bg1419", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj", "P_bg1420", "P_bg1414", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1418", "P_bg1419", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.30.ln_1", "P_bg1419"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1419", "P_bg1423", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1423", "P_bg1424", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.30.ln_2", "P_bg1424"),
 
-            **dense2("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj", "P_bg1424", "P_bg1425", bias=True),
+            **dense2("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj", "P_bg1426", "P_bg1427", bias=True),
             **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1424", "P_bg1425", bias=True),
             **norm("conditioner.embedders.1.model.transformer.resblocks.31.ln_1", "P_bg1425"),
             **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1425", "P_bg1429", bias=True),
